Refactor virt models in their respective namespaces
